requirements-prod.txt
# 生产环境依赖
# 基础依赖
-e .
alembic==1.16.5
apscheduler==3.11.0
bcrypt==4.3.0
bleach==6.2.0
blinker==1.9.0
cachelib==0.13.0
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.1
distlib==0.4.0
filelock==3.19.1
flask==3.1.2
flask-bcrypt==1.0.1
flask-caching==2.3.1
flask-cors==6.0.1
flask-jwt-extended==4.7.1
flask-login==0.6.3
flask-migrate==4.1.0
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
greenlet==3.2.4
itsdangerous==2.2.0
jinja2==3.1.6
kombu==5.5.4
loguru==0.7.3
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.2
mdurl==0.1.2
mypy==1.18.1
mypy-extensions==1.1.0
nodeenv==1.9.1
oracledb==3.3.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pre-commit==4.3.0
prompt-toolkit==3.0.52
psutil==7.1.0
psycopg==3.2.10
psycopg-binary==3.2.10
pycparser==2.23
pygments==2.19.2
pyjwt==2.10.1
pymssql==2.3.7
pymysql==1.1.2
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
redis==6.4.0
requests==2.32.5
rich==14.1.0
ruff==0.13.0
six==1.17.0
sqlalchemy==2.0.43
stevedore==5.5.0
structlog==25.4.0
types-pyyaml==6.0.12.20250915
typing-extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
virtualenv==20.34.0
wcwidth==0.2.13
webencodings==0.5.1
werkzeug==3.1.3
win32-setctime==1.2.0
wtforms==3.2.1

# 生产环境WSGI服务器
gunicorn==21.2.0
gevent==23.9.1

# 进程管理
supervisor==4.2.5

# 监控和日志
prometheus-client==0.19.0
